Calculo vel
Se incluye el calculo de las velocidades que tomarán los motores en la trayectoria, incluyendo el decrecimiento de ésta según se acerca al objetivo

+++ Motor.cpp
@@ -41,7 +41,7 @@ float motor::getFeedback(){
   }
 }
 
-void motor::setVelocidad(int a){
+void motor::setVelocidad(float a){
   velocidad = a;
   analogWrite(pin_pwm, velocidad);
 }
